requirements.txt
matplotlib
plotly
streamlit
langchain
langchain-openai
openai
python-dotenv
bs4
unstructured
tiktoken
loguru
watchdog